+++ AppCenter/AppCenter/include/MSAppCenterErrors.h
@@ -1,0 +1,1 @@
+../MSAppCenterErrors.h