+++ AppCenterAnalytics/AppCenterAnalytics/include/MSAnalyticsAuthenticationProvider.h
@@ -1,0 +1,1 @@
+../TransmissionTarget/MSAnalyticsAuthenticationProvider.h