+++ AppCenterAnalytics/AppCenterAnalytics/include/MSAnalyticsTransmissionTarget.h
@@ -1,0 +1,1 @@
+../TransmissionTarget/MSAnalyticsTransmissionTarget.h