+++ AppCenterAnalytics/AppCenterAnalytics/include/MSEventLog.h
@@ -1,0 +1,1 @@
+../Model/MSEventLog.h