+++ AppCenterAnalytics/AppCenterAnalytics/include/MSEventProperties.h
@@ -1,0 +1,1 @@
+../Model/MSEventProperties.h