+++ AppCenterCrashes/AppCenterCrashes/include/MSErrorAttachmentLog.h
@@ -1,0 +1,1 @@
+../Model/MSErrorAttachmentLog.h