+++ Scripts/build-ios-framework.sh
@@ -22,6 +22,9 @@ echo "Building ${TARGET_NAME}."
 # The following line create it in the root folder of the current project.
 PRODUCTS_DIR="${SRCROOT}/../AppCenter-SDK-Apple/iOS"
 
+# The directory to gather all frameworks and build it into xcframework.
+XCFRAMEWORK_DIR="${SRCROOT}/../AppCenter-SDK-Apple/xcframework"
+
 # Working dir will be deleted after the framework creation.
 WORK_DIR=build
 DEVICE_DIR="${WORK_DIR}/Release-iphoneos/"
@@ -47,6 +50,11 @@ mkdir -p "${PRODUCTS_DIR}"
 
 # Copy framework.
 cp -R "${DEVICE_DIR}/${PROJECT_NAME}.framework" "${PRODUCTS_DIR}"
+
+mkdir -p "${XCFRAMEWORK_DIR}"
+
+# Copy all framework files to use them for xcframework file creation.
+cp -R "${WORK_DIR}/" "${XCFRAMEWORK_DIR}"
 
 # Copy the resource bundle for App Center Distribute.
 if [ -d "${SRCROOT}/${DEVICE_DIR}/${RESOURCE_BUNDLE}.bundle" ]; then