+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### App Center
 
 * **[Improvement]** Use namespaced `NSUserDefaults` keys with the **MSAppCenter** prefix for all the keys set by the SDK. Fixed a few keys missing namespace.
+* **[Feature]** Partial Swift Package Manager distribution support added (for Core and Analytics modules).
+
 ___
 
 ## Version 3.1.1