+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 * **[Improvement]** Use namespaced `NSUserDefaults` keys with the **MSAppCenter** prefix for all the keys set by the SDK. Fixed a few keys missing namespace.
 * **[Feature]** Swift Package Manager support for Analytics and Crashes modules.
 
+### App Center Crashes
+
+* **[Improvement]** Update PLCrashReporter to 1.6.0.
+
 ___
 
 ## Version 3.1.1